requirements/automated-documentation-tests.txt
-r requirements.txt
ipykernel
ipywidgets
nbsphinx

numpydoc
pillow
pygments >= 2.4.1
pytest
setuptools-scm
sphinx <= 2.4.4
sphinx-automodapi
sphinx-copybutton
sphinx-gallery
sphinx_rtd_theme
nbconvert<6.0.0